backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
pymysql==1.1.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
b2sdk==2.7.0
ebooklib==0.18
lxml==5.3.0
httpx==0.28.1
pydantic-settings==2.7.1
Pillow==11.1.0
bleach==6.2.0
